requirements.txt
aiohttp==3.4.4
astroid==2.0.4
async-timeout==3.0.0
attrs==18.2.0
Cerberus==1.2
chardet==3.0.4
flatten-dict==0.0.3.post1
idna==2.7
idna-ssl==1.1.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
motor==2.0.0
multidict==4.4.2
pathlib2==2.3.2
PyJWT==1.6.4
pylint==2.1.1
pymongo==3.7.1
six==1.11.0
typed-ast==1.1.0
watchgod==0.2
wrapt==1.10.11
yarl==1.2.6
